requirements.txt
# Elasticsearch客户端
elasticsearch==8.6.2

# LLM和Agent框架
langchain-core>=0.1.0
langchain-openai>=0.0.1
langchain>=0.1.0
langgraph>=0.0.30

# 数据处理
pandas>=1.5.0
openpyxl>=3.10.0

# HTTP请求
requests>=2.28.0

# 环境变量管理
python-dotenv>=0.20.0